requirements.txt
fastapi
uvicorn
bcrypt
aiohttp
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
